+++ packages/compiler-cli/index.ts
@@ -8,10 +8,6 @@
 import {NodeJSFileSystem, setFileSystem} from './src/ngtsc/file_system';
 
 export {AotCompilerHost, AotCompilerHost as StaticReflectorHost, StaticReflector, StaticSymbol} from '@angular/compiler';
-export {DiagnosticTemplateInfo, getExpressionScope, getTemplateExpressionDiagnostics} from './src/diagnostics/expression_diagnostics';
-export {AstType, ExpressionDiagnosticsContext} from './src/diagnostics/expression_type';
-export {BuiltinType, DeclarationKind, Definition, PipeInfo, Pipes, Signature, Span, Symbol, SymbolDeclaration, SymbolQuery, SymbolTable} from './src/diagnostics/symbols';
-export {getClassMembersFromDeclaration, getPipesTable, getSymbolQuery} from './src/diagnostics/typescript_symbols';
 export {VERSION} from './src/version';
 
 export * from './src/metadata';
