+++ packages/compiler-cli/test/diagnostics/BUILD.bazel
@@ -1,28 +1,11 @@
 load("//tools:defaults.bzl", "jasmine_node_test", "ts_library")
 
-ts_library(
-    name = "mocks",
-    testonly = True,
-    srcs = [
-        "mocks.ts",
-    ],
-    deps = [
-        "//packages:types",
-        "//packages/compiler",
-        "//packages/compiler-cli",
-        "//packages/compiler-cli/test:test_utils",
-        "//packages/core",
-        "@npm//typescript",
-    ],
-)
-
 # check_types_spec
 ts_library(
     name = "check_types_lib",
     testonly = True,
     srcs = ["check_types_spec.ts"],
     deps = [
-        ":mocks",
         "//packages/compiler-cli",
         "//packages/compiler-cli/test:test_utils",
         "@npm//typescript",
@@ -44,69 +27,6 @@ jasmine_node_test(
     ],
     deps = [
         ":check_types_lib",
-        "//packages/core",
-        "//tools/testing:node",
-    ],
-)
-
-# expression_diagnostics_spec
-ts_library(
-    name = "expression_diagnostics_lib",
-    testonly = True,
-    srcs = ["expression_diagnostics_spec.ts"],
-    deps = [
-        ":mocks",
-        "//packages/compiler",
-        "//packages/compiler-cli",
-        "//packages/compiler-cli/test:test_utils",
-        "//packages/language-service",
-        "@npm//typescript",
-    ],
-)
-
-jasmine_node_test(
-    name = "expression_diagnostics",
-    bootstrap = ["angular/tools/testing/init_node_spec.js"],
-    data = [
-        "//packages/common:npm_package",
-        "//packages/core:npm_package",
-        "//packages/forms:npm_package",
-    ],
-    tags = [
-        # Disabled as these tests pertain to diagnostics in the old ngc compiler. The Ivy ngtsc
-        # compiler has its own tests for diagnostics.
-        "no-ivy-aot",
-    ],
-    deps = [
-        ":expression_diagnostics_lib",
-        "//packages/core",
-        "//tools/testing:node",
-    ],
-)
-
-# typescript_symbols_spec
-ts_library(
-    name = "typescript_symbols_lib",
-    testonly = True,
-    srcs = ["typescript_symbols_spec.ts"],
-    deps = [
-        ":mocks",
-        "//packages/compiler",
-        "//packages/compiler-cli",
-        "//packages/compiler-cli/test:test_utils",
-        "//packages/compiler/test:test_utils",
-        "//packages/language-service",
-        "@npm//typescript",
-    ],
-)
-
-jasmine_node_test(
-    name = "typescript_symbols",
-    bootstrap = ["angular/tools/testing/init_node_spec.js"],
-    data = [
-    ],
-    deps = [
-        ":typescript_symbols_lib",
         "//packages/core",
         "//tools/testing:node",
     ],
