+++ packages/language-service/src/symbols.ts
