+++ packages/language-service/src/types.ts
@@ -7,9 +7,9 @@
  */
 
 import {CompileDirectiveMetadata, NgAnalyzedModules, StaticSymbol} from '@angular/compiler';
-import {BuiltinType, DeclarationKind, Definition, PipeInfo, Pipes, Signature, Span, Symbol, SymbolDeclaration, SymbolQuery, SymbolTable} from '@angular/compiler-cli/src/language_services';
 
 import {AstResult} from './common';
+import {BuiltinType, DeclarationKind, Definition, PipeInfo, Pipes, Signature, Span, Symbol, SymbolDeclaration, SymbolQuery, SymbolTable} from './symbols';
 
 export {
   BuiltinType,
@@ -25,6 +25,7 @@ export {
   SymbolQuery,
   SymbolTable
 };
+
 
 /**
  * The information `LanguageService` needs from the `LanguageServiceHost` to describe the content of
